Created dotfiles script skeleton
Added a file called dotfiles. It is a skeleton for a perl script, which will be used for syncing the files across devices. Currently it uses only base imports, to prevent the need to download any extras except perl installed on the destination device. Only flag resolution and default path variables have been coded.